refactor: glyph nms logic into GlyphHandler nms classes
