+++ core/src/main/java/io/th0rgal/oraxen/font/FontManager.java
@@ -65,11 +65,8 @@ public class FontManager {
             loadFonts(fontConfiguration.getConfigurationSection("fonts"));
 
         useNmsGlyphs = GlyphHandlers.isNms() && NMSHandlers.getHandler() != null;
-        if (VersionUtil.atOrAbove("1.20.5") && useNmsGlyphs) {
-            Logs.logError("Oraxens NMS Glyph system is not working for 1.20.5...");
-            useNmsGlyphs = false;
-        } else if (useNmsGlyphs) {
-            NMSHandlers.getHandler().setupNmsGlyphs();
+        if (useNmsGlyphs) {
+            NMSHandlers.getHandler().glyphHandler().setupNmsGlyphs();
             Logs.logSuccess("Oraxens NMS Glyph system has been enabled!");
             Logs.logInfo("Disabling packet-based glyph systems", true);
             OraxenPlugin.get().getProtocolManager().removePacketListener(new InventoryPacketListener());
